+++ tests/test_misc_stuff.py
@@ -1,10 +1,13 @@
 import shutil
 import tempfile
 import uuid
+from json import dumps
+from json import loads
 from pathlib import Path
 
 import pytest
 from httpx import HTTPStatusError
+from httpx import Request
 from httpx import codes
 from pytest_httpx import HTTPXMock
 
@@ -105,3 +108,71 @@ class TestServerErrorRetry:
             with pytest.raises(HTTPStatusError) as exc_info:
                 _ = route.index(test_file).run_with_retry(initial_retry_wait=0.1, retry_scale=0.1)
             assert exc_info.value.response.status_code == codes.NOT_FOUND
+
+
+class TestWebhookHeaders:
+    def test_webhook_basic_headers(self, client: GotenbergClient, httpx_mock: HTTPXMock):
+        httpx_mock.add_response(method="POST", status_code=codes.OK)
+
+        client.add_webhook_url("http://myapi:3000/on-success")
+        client.add_error_webhook_url("http://myapi:3000/on-error")
+
+        test_file = SAMPLE_DIR / "basic.html"
+        with client.chromium.html_to_pdf() as route:
+            _ = route.index(test_file).run_with_retry()
+
+        requests = httpx_mock.get_requests()
+
+        assert len(requests) == 1
+
+        request: Request = requests[0]
+
+        assert "Gotenberg-Webhook-Url" in request.headers
+        assert request.headers["Gotenberg-Webhook-Url"] == "http://myapi:3000/on-success"
+        assert "Gotenberg-Webhook-Error-Url" in request.headers
+        assert request.headers["Gotenberg-Webhook-Error-Url"] == "http://myapi:3000/on-error"
+
+    def test_webhook_http_methods(self, client: GotenbergClient, httpx_mock: HTTPXMock):
+        httpx_mock.add_response(method="POST", status_code=codes.OK)
+
+        client.add_webhook_url("http://myapi:3000/on-success")
+        client.set_webhook_http_method("POST")
+        client.add_error_webhook_url("http://myapi:3000/on-error")
+        client.set_error_webhook_http_method("GET")
+
+        test_file = SAMPLE_DIR / "basic.html"
+        with client.chromium.html_to_pdf() as route:
+            _ = route.index(test_file).run_with_retry()
+
+        requests = httpx_mock.get_requests()
+
+        assert len(requests) == 1
+
+        request: Request = requests[0]
+
+        assert "Gotenberg-Webhook-Method" in request.headers
+        assert request.headers["Gotenberg-Webhook-Method"] == "POST"
+        assert "Gotenberg-Webhook-Error-Method" in request.headers
+        assert request.headers["Gotenberg-Webhook-Error-Method"] == "GET"
+
+    def test_webhook_extra_headers(self, client: GotenbergClient, httpx_mock: HTTPXMock):
+        httpx_mock.add_response(method="POST", status_code=codes.OK)
+
+        headers = {"Token": "mytokenvalue"}
+        headers_str = dumps(headers)
+
+        client.set_webhook_extra_headers(headers)
+
+        test_file = SAMPLE_DIR / "basic.html"
+        with client.chromium.html_to_pdf() as route:
+            _ = route.index(test_file).run_with_retry()
+
+        requests = httpx_mock.get_requests()
+
+        assert len(requests) == 1
+
+        request: Request = requests[0]
+
+        assert "Gotenberg-Webhook-Extra-Http-Headers" in request.headers
+        assert request.headers["Gotenberg-Webhook-Extra-Http-Headers"] == headers_str
+        assert loads(request.headers["Gotenberg-Webhook-Extra-Http-Headers"]) == headers